requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
colorama==0.4.6
fastapi==0.109.0
h11==0.14.0
httptools==0.6.1
idna==3.6
pydantic==2.5.3
pydantic[email]
pydantic_core==2.14.6
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.25.0
watchfiles==0.21.0
websockets==12.0
SQLAlchemy==1.4.45
envparse==0.2.0
psycopg2==2.9.5
asyncpg==0.27.0
alembic==1.9.0
pytest==7.4.4
httpx==0.26.0
pytest-asyncio==0.23.3
pre-commit==3.6.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2

